refactor: improved code structure and removed unused files
