+++ src/features/vibes/styles/UserVibes.css
@@ -7,6 +7,7 @@
   min-height: 180px;
   cursor: pointer;
   opacity: 1;
+  border: 1px solid rgba(0, 0, 0, 0.05);
   transition:
     transform 0.18s cubic-bezier(.48,.07,.75,.95),
     box-shadow 0.20s cubic-bezier(.48,.07,.75,.95),
