+++ src/shared/ui/Footer.css
@@ -1,6 +1,6 @@
 .footer {
-  background-color: #f9f9f9; /* ← светлый фон */
-  color: #333;              /* ← тёмный текст */
+  background-color: #f9f9f9; 
+  color: #333;              
   padding: 2rem 1rem;
   text-align: center;
   font-size: 0.85rem;
