+++ src/shared/ui/Header.css
@@ -13,7 +13,7 @@
 
 /* ------------------------------HEADER-------------------------- */
 .header {
-  max-width: 1280px; /*Just like it used to be*/
+  max-width: 1280px; 
   margin: 1.5rem auto 0 auto;
   width: 100%;
   background-color: #fff;
@@ -27,12 +27,12 @@
 /* ------------------------------HEADER'S LINKS-------------------- */
 .header__container {
   max-width: 1100px;
-  height: 85px; /* header's height */
+  height: 85px; 
   margin: 0 auto;
   display: flex;
-  align-items: center; /* center vertically */
+  align-items: center; 
   justify-content: space-between;
-  padding: 0 1rem; /* horizontal indents */
+  padding: 0 1rem; 
 }
 
 /* -------------------------------LOGO------------------------------ */
@@ -47,7 +47,7 @@
   height: 20px;
 }
 .logo img {
-  width: 28px; /* уменьшил размер иконки */
+  width: 28px; 
   height: 28px;
   object-fit: contain;
   display: block;
@@ -83,7 +83,7 @@
   text-decoration: underline;
 }
 .early-access-bubble {
-  background-color: #ffeaea; /* light red background */
+  background-color: #ffeaea; 
   border: 2px solid red;
   border-radius: 8px;
   padding: 12px 16px;
@@ -98,7 +98,6 @@
 }
 
 .early-access-bubble::after {
-  /* remove tail */
   content: none;
 }
 
@@ -125,9 +124,9 @@
 }
 
 .subscribed-green {
-  background-color: #e6f9e6; /* light green background */
+  background-color: #e6f9e6; 
   border: 2px solid #4caf50;
-  color: #2e7d32; /* dark green text */
+  color: #2e7d32; 
 }
 
 .subscribed-green .early-access-text,
